membuat method get all dan delete di user

+++ src/main/java/com/Binusa/BawasluServer/repository/UserRepository.java
@@ -5,9 +5,9 @@ package com.Binusa.BawasluServer.repository;
 import com.Binusa.BawasluServer.model.UserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
+// UserRepository.java
 @Repository
-public interface UserRepository extends JpaRepository<UserModel, Integer> {
+public interface UserRepository extends JpaRepository<UserModel, Long> {
     UserModel findByUsername(String username);
 
     UserModel findById(long userId);
